--- a/src/Datadog.Trace/SpanContextPropagator.cs
+++ b/src/Datadog.Trace/SpanContextPropagator.cs
@@ -33,7 +33,9 @@ namespace Datadog.Trace
         /// </summary>
         /// <param name="context">A <see cref="SpanContext"/> value that will be propagated into <paramref name="headers"/>.</param>
         /// <param name="headers">A <see cref="IHeadersCollection"/> to add new headers to.</param>
-        public void Inject(SpanContext context, IHeadersCollection headers)
+        /// <typeparam name="T">Type of header collection</typeparam>
+        public void Inject<T>(SpanContext context, T headers)
+            where T : IHeadersCollection
         {
             if (context == null) { throw new ArgumentNullException(nameof(context)); }