- Make SpanContextPropagator.Inject generic to prevent the boxing of the headers collection
- Add an overload in TypeExtensions to avoid building the full type name
- Stop using String keys in AsyncHelper
- Reuse the boxed cancellationtoken when needed
- Split SendAsyncInternal to avoid allocating the async state machine on the fast path
- Cache the "namespace and name filters" array
- Cache the return task type

**Benchmark:**

```csharp
public class CustomMessageHandler : HttpMessageHandler
{
	private static Task<HttpResponseMessage> _cachedResult = Task.FromResult(new HttpResponseMessage());
	private static HttpResponseMessage _cachedResponse = new HttpResponseMessage();

	protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
	{
		return _cachedResult;
	}
}

(Task)HttpMessageHandlerIntegration.HttpMessageHandler_SendAsync(...).Wait();
```

FastPath: Takes the "skip instrumentation" path in `SendAsyncInternal`
Full: Goes through the instrumentation code

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-VTGAFA : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-MNUBEA : .NET Core 3.1.6 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.31603), X64 RyuJIT


```

**.NET 4.7.2:**

|                    Method |          Mean |         Error |         StdDev |        Median |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|-------------------------- |--------------:|--------------:|---------------:|--------------:|-------:|-------:|-------:|----------:|
| SendRequest_NoIntegration |      3.781 ns |     0.1032 ns |      0.1447 ns |      3.706 ns |      - |      - |      - |         - |
|      SendRequest_Full_Old | 18,710.139 ns |   399.8844 ns |    460.5075 ns | 18,551.764 ns | 3.2043 | 0.4578 | 0.1526 |   20201 B |
|      SendRequest_Full_New | 11,077.879 ns |   121.0923 ns |    107.3451 ns | 11,080.820 ns | 0.4120 | 0.1984 |      - |    2624 B |
|  SendRequest_Fastpath_Old |  6,020.003 ns |    94.4310 ns |     88.3308 ns |  5,999.944 ns | 2.8534 | 0.0153 |      - |   17991 B |
|  SendRequest_Fastpath_New |  2,177.383 ns |    42.6974 ns |     37.8501 ns |  2,162.498 ns | 0.0725 | 0.0076 |      - |     457 B |


**.NET Core 3.1:**

|                    Method |          Mean |         Error |         StdDev |        Median |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|-------------------------- |--------------:|--------------:|---------------:|--------------:|-------:|-------:|-------:|----------:|
| SendRequest_NoIntegration |      4.122 ns |     0.0380 ns |      0.0337 ns |      4.116 ns |      - |      - |      - |         - |
|      SendRequest_Full_Old | 15,527.897 ns |   304.9692 ns |    501.0731 ns | 15,470.544 ns | 3.4485 | 0.5798 | 0.1526 |   21319 B |
|      SendRequest_Full_New |  7,810.876 ns |   151.9542 ns |    134.7034 ns |  7,782.911 ns | 0.3815 | 0.1831 |      - |    2416 B |
|  SendRequest_Fastpath_Old |  5,331.927 ns |   105.6334 ns |    187.7632 ns |  5,289.986 ns | 3.0823 | 0.0153 |      - |   19359 B |
|  SendRequest_Fastpath_New |  1,132.209 ns |     6.6264 ns |      5.5334 ns |  1,131.997 ns | 0.0725 |      - |      - |     456 B |