@kevingosse 2020-07-16
It's already a nice improvement but I believe there is still room for more. I'll iterate on this, either in this PR or in a separate one.