@kevingosse 2020-07-17
In the end I stopped relying on AsyncHelper and added an extension method to convert a task type to another. It makes the code much easier to follow (in my opinion).

I also moved the "no instrumentation" path earlier in the method to reduce the overhead.

Final results:

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.1.302
  [Host]     : .NET Core 3.1.6 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.31603), X64 RyuJIT
  Job-UDBQJC : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-UEXFKP : .NET Core 3.1.6 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.31603), X64 RyuJIT


```
|                   Method |       Runtime |        Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------------- |-------------- |------------:|----------:|----------:|-------:|-------:|------:|----------:|
|     SendRequest_Full_Old |    .NET 4.7.2 | 14,343.3 ns | 410.78 ns | 403.44 ns | 3.1738 |      - |     - |   20115 B |
|     SendRequest_Full_New |    .NET 4.7.2 |  9,191.7 ns |  76.87 ns |  68.14 ns | 0.3815 |      - |     - |    2423 B |
| SendRequest_Fastpath_Old |    .NET 4.7.2 |  5,445.0 ns |  31.44 ns |  29.41 ns | 2.8687 | 0.0153 |     - |   18072 B |
| SendRequest_Fastpath_New |    .NET 4.7.2 |  1,321.7 ns |  15.28 ns |  13.54 ns | 0.0343 |      - |     - |     217 B |
|     SendRequest_Full_Old | .NET Core 3.1 | 12,221.1 ns |  67.97 ns |  60.26 ns | 3.3875 | 0.0153 |     - |   21295 B |
|     SendRequest_Full_New | .NET Core 3.1 |  6,074.7 ns |  56.31 ns |  49.91 ns | 0.3510 |      - |     - |    2224 B |
| SendRequest_Fastpath_Old | .NET Core 3.1 |  5,439.5 ns | 107.52 ns |  95.31 ns | 3.0899 | 0.0153 |     - |   19431 B |
| SendRequest_Fastpath_New | .NET Core 3.1 |    563.3 ns |   9.63 ns |   8.54 ns | 0.0343 |      - |     - |     216 B |